brew: module was dropping task initializations by mistake
In previous patches, parameters like `task_ids` were wrapped within
`kwargs`, and as such passed to parent's implementation to handle them.
With addition of type annotations, these parameters were unpacked form
`kwargs` but the call to parent was *not* updated to reflect the fact
they are no longer part of `kwargs`. THis effectively dropped any
initialization but `build_ids`.

Resolves TFT-844