Generate reproducer commands for tmt
We want to show user a nice way how he can reproduce problems
via his localhost. For that we need a set of commands for each
plan, where:

* git clone commands clone the tested repo
* enviornment variables file is downloaded from Testing Farm
* a tmt command exists which works on localhost

The command is gathered as list of strings inside a schedule entry,
as each reproducer is specific to a schedule entry (i.e. one plan
run in our case)

Relates to TFT-641

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>