pipeline-state-reporter: support saving final msg to file
While working on support for gathering errors from Jenkins jobs
I realized it would be really handy to have the final message
around in the artifacts, so I can easily parse it.

So I decided extend `pipeline-state-reporter` to save the final
message to a file. Other messages are around in the citool debug
log, we really need just this one, and just the body.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>