install_copr_build: Simplify installation
Drop the redundant `yum/dnf update` steps. They do nothing in addition
to the already existing install/downgrade/reinstall steps.

We call `dnf install` with a specific set of RPM URLs. It will happily
upgrade/downgrade any existing RPM without having to do `downgrade`
explicitly. Drop it for dnf. This doesn't work with yum on RHEL 7,
though: `yum install` refuses to downgrade, so there we still need the
explicit `yum downgrade` step.  Don't change the failure behaviour
there, as that would be too intrusive for the code.

This has the added benefit that we can strengthen the commands to
require that the `dnf install` step actually succeeds -- then package
installation failures will not be silently ignored any more (which will
probably cause the test to fail later on, but with a less obvious
message).  The only exception is `dnf reinstall`, which behaves really
nasty, so that still needs to be `|| true`'d.  Add reference to
https://bugzilla.redhat.com/show_bug.cgi?id=1831022 for tracking this.