Add `debug` module
When I work with pipelines, I often would like to peak
into it once it is finishing or when it fails. I used
to insert debugger into the code for that, but in some
cases a simple module calling ipdb can be useful.

So I wrote this simple module and hope you might like
it also.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>