test_schedule_tmt: Add unit test for results XML generation
In particular, check that it picks up tmt's logs as expected. This
currently only looks at the first log of tmt's results.yaml. That will
be extended later, but capture the status quo for now.

This has to add some necessary setup and mocks for `tmt run` and guest.
These are already present in the `testing-farm` branch for other unit
tests, and thus will be added eventually anyway.