test_schedule_tmt: Drop log sorting
Instead, add the logs in the order that we want. It makes some sense to
start with the artifacts directory, and either way that is kind of API
now (i.e. how it behaved before). But conceptually this shouldn't be
sorted alphabetically but by relevancy. Even when adding arbitrary
additional logs, the first one should always be the primary output that
developers should look at.