coldstore: handle artifacts location containing '#'
While digging into why URLs are broken for tests which
contain `TC#12345` like strings, we found that
`urlnormalizer` python package treats it as a URL
fragement and drops everything after it. As that seems
an actually valid thing to do for normalizing the URL,
do not let '#' even enter the normalizer and handle
the quoting for this special use case.

Add tests for proof.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>