install_koji_build: Simplify installation
Drop the no-op `yum/dnf update` steps.

We call `dnf install` with a specific set of RPM URLs. It will happily
upgrade/downgrade any existing RPM without having to do `downgrade`
explicitly. Drop it for dnf. This doesn't work with yum on RHEL 7,
though: `yum install` refuses to downgrade, so there we still need the
explicit `yum downgrade` step.  Don't change the failure behaviour
there, as that would be too intrusive for the code.

But stop ignoring exceptions for the dnf case. This will make it more
robust to record the packages and show them to the user as reproducer,
so that they can work in a `set -e` context (such as tmt's
`prepare --how shell`).

This is essentially the same as what commit f5082a51bcf has done for
install_copr_build.