artemis: retry on connection error
We recently made the exceptions handling only expected exceptions.

And suprise, seems for a very long time we were ignoring a random
DNS resolution problem lurking there.

As there seems there is a bug somewhere in podman causing random connection
errors, plus the things for DNS got a lot worse with introducing `aadvardk-dns`
seems with retries here and removing that package we can have the kind
of "stability" as before.

Also to get the scope of this problem, we will now have sentry log
entries.

I will continue now contacting podman to get this networking stack
problem fixed, as it is lurking there, and I believe it is also
causing random failures users sometime see.

Related to TFT-1755

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>