test_schedule_tmt: Fix collecting SUT install commands
Commit b7bc29426 picked up an obsolete way of reading the install
commands via a shared function. But commit e0f47392c01be2 changed that
to writing to a file. Adjust writing tmt-reproducer.sh to read the file
instead, like in the testing-farm branch.

Port the COPR and koji TMT scheduler tests from the testing-farm branch
to exercise that code path. Drop the bogus `sut_install_commands`
scenario from `test_tmt_output_dir()`, as that was testing mocked
functionality which does not exist in the real code.