Resolve race condition when installing copr builds
As it turns out, the code needs the copr builds initialization
to be thread safe.

But the currently used `tasks` function is not thread safe.

Add a thread safe variant for it, named it `copr_builds`, as it
is specific to copr builds only.

And use this function instead, to resolve TFT-1970.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>