git: unify checkout_ref, support commits from refs
In case of merging, it can be expected the commit comes from a refspec.

After investigating, it is more reasonable to just use one way of
checking out refs, instead of the previous 2 methods.

While this makes the checkout a bit longer, it makes the code more
simple, having just one method. Moreover the problem if commit comes
from a refspec works.

Together with that, we need to modify the default refspec for `git fetch`, support
it supports also Pagure and GitHub. Since we can specify multiple rules in the config,
just configure them all at once instead of “patching” around based on the clone URL from the
request.

Signed-off-by: Matej Focko <mfocko@redhat.com>
Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>