pyproject.toml: limit upgrade to Ansible 2.14
After few hours of investigation of TFT-2072, the culprint
is the new Ansible 2.15 breaks STI.

We agreed on freezing to 2.14. We updated to to 2.15, because
we install the deps via pip, what has great but poetry does
not yet make it possible to freeze deps:

    https://github.com/python-poetry/poetry/issues/2778

This limits the requirements to ansible-core<2.15.

Resolves TFT-2072

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>