libs/sut_installation: fixed for dnf5
Well, `dnf5` is now keen on options, they cannot be global anymore.
As this is an expected breaking change, fix the `--allowerasing`
position in the command.

Note that `allowerasing` from `reinstall` command is temporarily
dropped until:

https://github.com/rpm-software-management/dnf5/pull/637

Is resolved in Fedora Rawhide.
I do believe it is a corner case, and we can survive with it
for few days to not work.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>