artemis: support setting workdir for storing runtime data
Currently artemis stores everything in the current directory where
gluetool runs. It would be good if it would be able to store runtime
data also in a working directory, for example the one schedule entry
has.

Introduce a `workdir` optional parameter to the `provision` method.
Next MR will do the test schedule runner part to pass the workdir to it.

Also cover the feature with tests, together with the console log.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>