artemis: improve destroying of guests
This patch fixes multiple issues related to destroying of the guests:

1. make sure an unfinished Artemis guest is correctly cancelled
2. leave the destroying of leftover guests (i.e. guests not cancelled
   from test schedule modules) during module destroy

To be able to achieve 1. we had to split the provisioning into two
parts:

* start guest provisioning
* wait for the provisioning to finishe

The guest is kept after 1. to be able to destroy it.

For the 2. remove the rogue `guest.destroy` and leave the cleanup to
destroy function to make the code more clean.

Update the tests to test correctly for everything.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>