tmt_test_schedule_tmt*: pass env vars to discover
The environment variables need to be passed also to discover, so cases
when users use variables in discover phase work.

Unfortunately a bit of refactoring was needed dur to this, the
`_apply_test_filter` function was changed to return True or False and
indicate if the plan should be added.

Resolves TFT-2192

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>