archive: support verification of archiving target location
Adding simple way to verify if the archivation target is reachable via
its artifacts location URL. Just add `verify: "true"` to the archived
item. For directories only the directory location is verified.

Requires for the `coldstore` module to come before archive module if
`verify` used in `progress` or `execute` stage. We plan to use it only
in `destroy` stage for now.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>