libs/xunit: redact long output from printable representation
In the `xunit` the system_out can be very long.
Lately this caused issues where the logs exploaded, as the results
are now part of the evaluation context. When they were written
out in verbose log, this representation leaked the whole output
there during each evaluation.

This should redact the possible fields with huge content from the
printable representation.

Related to TFT-2495

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>