tests/test_test_schedule_sti.py: fix logging
Without topics this code worked, but with added topics
we need to make sure to well initialize the TestScheduleEntry
which should be passed to `gather_test_results`. The original
code passed directly logger, what caused issues topic being
passed over to the main logging debug function, which was
not able to survive `topic` kwarg.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>