Correctly set error state for out-of-memory events
As I found out while testing on dev, the request is not transitioned
to the `error` state when the out-of-memory event happens.

That is because the pipeline `state` is not set when pipeline is canceled,
what is needed for cancellation events, but for oom events it causes the
state to stay in `running`.

Introduce and `oom_message` shared function which indicates
out-of-memory event happen when set and correctly the state to error in
that cases and also correctly set the summary.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>