oom: move to detection from psutil
When run in a container, count the rss memory of all processes
and use that for checking.

As it turned out, cgroups does not work well with nested containers
and reports inaccurate results, causing jobs to be terminated
unexpectedly. This should be a lot more accurate.

Note that shared memory used can be counted twice, but I believe it is not
an issue for our tooling, we do not use multiprocessing, and the run
`podman` is reported correctly.

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>