tmt_test_schedule_tmt*: add only `testout.log` to system_out
The tmt-1.49 version introduced `failures.yaml` file, which caused
the integration tests to fail, as they were added now to `system_out`
in the "native" xunit output.

I got fed up accepting this, and also this was utterly wrong for a long
time. Make sure we add only the test output log to `system_out`. Now
even binary files are processed, which we had issues before (solved by
limiting the system_out size.

With this patch only the a reasonable human output will get to the
`<system-out>` xunit element, as it should be. And of course, it fixes
the integration test expected output :)

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>