artemis: custom error generation from guest event log
With these changes we are able to provide nice error messages
for Artemis errors. The analysis of the errors can be done
by looking at last error events in the event log and some
Jinja2 template magic.

If the error template rendering fails, the original message is
returned.

The change contains an easy way to test the rendered error message
using:

```
$ citool artemis --analyze-event-log URL_TO_GUEST_EVENT_LOG
```

To test out:

* checkout this citool-config: https://gitlab.cee.redhat.com/baseos-qe/citool-config/-/merge_requests/15617
* try some of the errors:

```
$ citool artemis --analyze-event-log https://artifacts.osci.redhat.com/testing-farm/08505e02-214b-46fb-b677-980ca1f86ca3/work-test6od8l5dq/1bc3806f-3c49-4b5a-82f7-4d04aaccd522-artemis-guest-log.yaml
---v---v---v---v---v---
RHEL-7 product is not supported on aarch64.
---^---^---^---^---^---

$ citool artemis --analyze-event-log https://artifacts.osci.redhat.com/testing-farm/89327acc-517b-4ed1-ab96-a80e806f314e/work-bootc-fips-upgrade217uusz7/b3590845-6cf6-4ade-a597-a971b8a7aac7-artemis-guest-log.yaml
---v---v---v---v---v---
Failed to find image tf-build-rhel-10-0-updates-20251211-1-s390x in ibmcloud-us-south-3 pool.
Please file an issue to Testing Farm, this is a infrastructure problem that needs fixing by the maintainers.
---^---^---^---^---^---
```

Signed-off-by: Miroslav Vadkerti <mvadkert@redhat.com>